% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_gamma_R.R
\name{adaptive_gamma_ess}
\alias{adaptive_gamma_ess}
\title{ESS target adaptive gamma criterion
Binary search such that ess0 <= ESS <= ess1}
\usage{
adaptive_gamma_ess(R_evc, R_evl, C, n, ess0 = 0.5, ess1 = 0.5, imin = 0,
  imax = 40, delta_gam = 1/imax)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{C}{Yb'R^-1(y - Hxbbar)}

\item{n}{ensemble size}

\item{ess0}{ESS lower bound}

\item{ess1}{ESS upper bound}

\item{imax, }{imin: gammas considered go from imin/(imax-imin+1) to imax/(imax-imin+1),}

\item{delta_gam}{gammas increments, can be changed}
}
\value{
list(gamma, w, ess) of chosen model
}
\description{
ESS target adaptive gamma criterion
Binary search such that ess0 <= ESS <= ess1
}
