% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_gamma_R.R
\name{adaptive_gamma_minSCORE}
\alias{adaptive_gamma_minSCORE}
\title{Score-based adaptive gamma criterion
Grid search to minimize criterion}
\usage{
adaptive_gamma_minSCORE(R_evc, R_evl, C, n, imin = 0, imax = 40,
  delta_gam = 1/imax, unif = runif(1), score_type = "MSE",
  eps_cv = matrix(rnorm(n * n), n, n), ...)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{C}{Yb'R^-1(y - Hxbbar)}

\item{n}{ensemble size}

\item{imax, }{imin: gammas considered go from imin/(imax-imin+1) to imax/(imax-imin+1),}

\item{delta_gam}{gammas increments, can be changed}

\item{unif}{used in balanced sampling}

\item{score_type}{describe which score to use (MSE, ES, MSECV, ESCV)}

\item{eps_cv}{perturbations to pass for CV scheme}
}
\value{
list(gam_best, w_best, ess_best, gammas, errors), where the last two are used to debug/plot
}
\description{
Score-based adaptive gamma criterion
Grid search to minimize criterion
}
