% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_gamma_R.R
\name{cv_score}
\alias{cv_score}
\title{CV score of ES, MSE or EMSE (in ensemble space)}
\usage{
cv_score(gam, R_evc, rvl, Sfull, unif, eps_cv, score_type, n)
}
\arguments{
\item{gam}{default gam value if not adaptive}

\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{rvl}{eigenvalues of S, already floored at 0}

\item{unif}{used in balanced sampling}

\item{eps_cv}{perturbations to pass for CV scheme}

\item{score_type}{MSE or ES}

\item{n}{ensemble size}
}
\value{
estimated CV score
}
\description{
CV score of ES, MSE or EMSE (in ensemble space)
}
