% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_gamma_R.R
\name{energy_score_es}
\alias{energy_score_es}
\title{Energy score of yhat predicted by xahat, where hat is the projection onto orthogonal basis of ensemble space:}
\usage{
energy_score_es(Wmu, ind, wmubar, R_evc, Cvec, S, rvl, rvl_inv, del_pag, n)
}
\arguments{
\item{Wmu}{as returned by get_Wmu}

\item{ind}{vector of resampling indices}

\item{wmubar}{as returned by get_wmubar}

\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{S}{S=Yb' R^-1 Yb'}

\item{rvl}{eigenvalues of S, already floored at 0}

\item{rvl_inv}{inverse of rvl such that 1/0=0}

\item{del_pag}{eigenvalues of Pag: Pag = R_evc diag(del_pag) R_evc'}

\item{n}{ensemble size}
}
\value{
estimated energy score
}
\description{
Energy score of yhat predicted by xahat, where hat is the projection onto orthogonal basis of ensemble space:
}
