% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_main_R.R
\name{etkpf_analysis}
\alias{etkpf_analysis}
\title{Wrapper around etkpf_es
compute the ensemble space elements needed to call etkpf_es}
\usage{
etkpf_analysis(xb, y, H, R, etkpf_foo = etkpf_es, ...)
}
\arguments{
\item{xb}{the background ensemble}

\item{y}{the observations}

\item{H}{observation operator}

\item{R}{observation covariance}

\item{etkpf_foo}{function to compute the ensemble space analysis}

\item{...}{additional arguments to pass to etkpf_foo}
}
\value{
a list with xa and mod=model returned by etkpf_foo (Wmu,Weps, etc)
}
\description{
Wrapper around etkpf_es
compute the ensemble space elements needed to call etkpf_es
}
