% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_main_f90.R
\name{etkpf_es_f90}
\alias{etkpf_es_f90}
\title{ETKPF analysis in ensemble space (in fortran90)
compute the W matrices of the analysis of ETKPF}
\usage{
etkpf_es_f90(R_evc, R_evl, C, n, weps_type, adaptive, center = 1,
  unif = runif(1), gam = 0.5, eps = matrix(rnorm(n * n), n, n),
  eps_cv = matrix(rnorm(n * n), n, n), ess0 = 0.5, ess1 = 0.5, imin = 0,
  imax = 40, ...)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{C}{Yb'R^-1(y - Hxbbar)}

\item{n}{ensemble size}

\item{weps_type}{Weps algorithm (0: stochastic, 1: Riccati)}

\item{adaptive}{scheme to choose gamma:
0: not adaptive
1: in ess bounds
2: min MSE
3: min ES  
4: min CV(MSE)
5: min CV(ES)}

\item{center}{(0: not centered, 1: centered such that  xabar = xbbar + Xb wm and xa = xabar + Xb W)}

\item{unif}{used in balanced sampling}

\item{gam}{default gam value if not adaptive}

\item{eps}{perturbations to pass in case weps=0}

\item{eps_cv}{perturbations to pass for CV scheme}
}
\description{
ETKPF analysis in ensemble space (in fortran90)
compute the W matrices of the analysis of ETKPF
}
