% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{get_alpha}
\alias{get_alpha}
\title{Compute alphai weights}
\usage{
get_alpha(R_evc, R_evl, C, n, gam)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{C}{Yb'R^-1(y - Hxbbar)}

\item{n}{ensemble size}

\item{gam}{fixed gamma value}
}
\value{
alphai
}
\description{
Compute alphai weights
}
