% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{get_Weps_riccati}
\alias{get_Weps_riccati}
\title{Compute Weps for Riccati ETKPF}
\usage{
get_Weps_riccati(R_evc, R_evl, n, gam, Wmu, ind_resample, tol = 10^-9,
  maxit = 20)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{n}{ensemble size}

\item{gam}{fixed gamma value}

\item{Wmu}{Wmu as returned from get_Wmu}

\item{ind_resample}{vector of resampling indices}

\item{tol}{tolerance for stopping criterion}

\item{maxit}{maximum number of iteration of Newton's algorithm}
}
\value{
Weps
}
\description{
Compute Weps for Riccati ETKPF
}
