% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{get_Weps_stochastic}
\alias{get_Weps_stochastic}
\title{Compute Weps for stochastic EnKPF}
\usage{
get_Weps_stochastic(R_evc, R_evl, n, gam, eps)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{n}{ensemble size}

\item{gam}{fixed gamma value}

\item{eps}{perturbations to pass in case weps=0}
}
\value{
Weps
}
\description{
Compute Weps for stochastic EnKPF
}
