% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{get_Wmu}
\alias{get_Wmu}
\title{Compute Wmu}
\usage{
get_Wmu(R_evc, R_evl, n, gam)
}
\arguments{
\item{R_evc}{eigenvectors of S=Yb'R^-1 Yb}

\item{R_evl}{eigenvalues of S}

\item{n}{ensemble size}

\item{gam}{fixed gamma value}
}
\value{
Wmu
}
\description{
Wmu is defined such that mu = xbbar + Xb wmubar + Xb Wmu
}
