% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_gp.R
\name{gp_as_df}
\alias{gp_as_df}
\title{Transform the a Gaussian random field matrix into a data frame}
\usage{
gp_as_df(state, method = NULL)
}
\arguments{
\item{state}{can be both a vector (one sample) or an ensemble qxk}
}
\description{
Transform the a Gaussian random field matrix into a data frame
}
