% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_gp.R
\name{gp_ens0}
\alias{gp_ens0}
\title{Generate an ensemble of Gaussian random field with same mean and covariance}
\usage{
gp_ens0(k, gp_data)
}
\arguments{
\item{k}{ensemble size}

\item{gp_data}{a list returned by gp_simulate}
}
\description{
Generate an ensemble of Gaussian random field with same mean and covariance
}
