% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_gp.R
\name{gp_plot}
\alias{gp_plot}
\title{Plot the true field, the observations and an ensemble}
\usage{
gp_plot(gp_data, gp_ens, tit = "")
}
\arguments{
\item{gp_data}{a list returned by gp_simulate}

\item{gp_ens}{a ensemble matrix qxk}

\item{tit}{optional plot title}
}
\description{
Plot the true field, the observations and an ensemble
}
