% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{lyapunov}
\alias{lyapunov}
\title{Lyapunov equation solver (used by get_Weps_riccati):}
\usage{
lyapunov(C, D, E, n)
}
\arguments{
\item{C}{left multiplier (A' + X1)}

\item{D}{right multiplier (A + X1)}

\item{E}{rhs (Pag + X1 X1')}

\item{n}{ensemble size}
}
\value{
X
}
\description{
CX + XD = E
in Riccati Newton's step used to solve for X2 given X1:
 (A' + X1) X2  + X2  (A + X1) = (Pag + X1 X1')
}
