% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{msqrt}
\alias{msqrt}
\title{computes a symmetric matrix square root of a positive definite matrix
use eigenvalue decompostion}
\usage{
msqrt(A)
}
\arguments{
\item{A}{a symmetric matrix}
}
\value{
S such that S'S = A
}
\description{
computes a symmetric matrix square root of a positive definite matrix
use eigenvalue decompostion
}
