% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{reorder_ind}
\alias{reorder_ind}
\title{Rearrange indices such that they match 1:n}
\usage{
reorder_ind(ind, n)
}
\arguments{
\item{ind}{vector of resampling indices}

\item{n}{ensemble size}
}
\value{
permuted indices such that they match 1:n as much as possible
}
\description{
Slow but same as in fortran90 to test code...
}
