% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etkpf_util_R.R
\name{reorder_ind_fast}
\alias{reorder_ind_fast}
\title{Rearrange indices such that they match 1:n}
\usage{
reorder_ind_fast(ind_resample, n)
}
\arguments{
\item{ind_resample}{vector of resampling indices}

\item{n}{ensemble size}
}
\value{
permuted indices such that they match 1:n as much as possible
}
\description{
Rearrange indices such that they match 1:n
}
